+++ motan-core/src/main/java/com/weibo/api/motan/config/AbstractRefererConfig.java
@@ -17,7 +17,6 @@
 package com.weibo.api.motan.config;
 
 /**
- *
  * Abstract referer config.
  *
  * @author fishermen
@@ -34,7 +33,6 @@ public abstract class AbstractRefererConfig extends AbstractInterfaceConfig {
     protected String p99;
     protected String p999;
     protected String errorRate;
-    protected Boolean asyncInitConnection;
 
     public String getMean() {
         return mean;
@@ -75,13 +73,4 @@ public abstract class AbstractRefererConfig extends AbstractInterfaceConfig {
     public void setErrorRate(String errorRate) {
         this.errorRate = errorRate;
     }
-
-    public Boolean getAsyncInitConnection() {
-        return asyncInitConnection;
-    }
-
-    public void setAsyncInitConnection(Boolean asyncInitConnection) {
-        this.asyncInitConnection = asyncInitConnection;
-    }
-
 }