+++ motan-core/src/main/java/com/weibo/api/motan/protocol/support/ProtocolFilterDecorator.java
@@ -1,11 +1,11 @@
 /*
  * Copyright 2009-2016 Weibo, Inc.
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software distributed under the License
  * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
  * or implied. See the License for the specific language governing permissions and limitations under
@@ -19,18 +19,20 @@ import com.weibo.api.motan.common.URLParamType;
 import com.weibo.api.motan.core.extension.Activation;
 import com.weibo.api.motan.core.extension.ActivationComparator;
 import com.weibo.api.motan.core.extension.ExtensionLoader;
+import com.weibo.api.motan.core.extension.SpiMeta;
 import com.weibo.api.motan.exception.MotanErrorMsgConstant;
 import com.weibo.api.motan.exception.MotanFrameworkException;
 import com.weibo.api.motan.filter.Filter;
 import com.weibo.api.motan.filter.InitializableFilter;
 import com.weibo.api.motan.rpc.*;
 import com.weibo.api.motan.util.LoggerUtil;
+import com.weibo.api.motan.util.StringTools;
 import org.apache.commons.lang3.StringUtils;
 
 import java.lang.reflect.Method;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.List;
+import java.util.*;
+
+import static com.weibo.api.motan.common.MotanConstants.DISABLE_FILTER_PREFIX;
 
 /**
  * Decorate the protocol, to add more features.
@@ -197,17 +199,16 @@ public class ProtocolFilterDecorator implements Protocol {
      * 获取方式：
      * 1）先获取默认的filter列表；
      * 2）根据filter配置获取新的filters，并和默认的filter列表合并；
-     * 3）再根据一些其他配置判断是否需要增加其他filter，如根据accessLog进行判断，是否需要增加accesslog
+     * 3）如果filter配置中有'-'开头的filter name，表示disable某个filter，对应的filter不会装配
      * </pre>
      *
      * @param url
      * @param key
      * @return
      */
-    private List<Filter> getFilters(URL url, String key) {
-
+    protected List<Filter> getFilters(URL url, String key) {
         // load default filters
-        List<Filter> filters = new ArrayList<Filter>();
+        List<Filter> filters = new ArrayList<>();
         List<Filter> defaultFilters = ExtensionLoader.getExtensionLoader(Filter.class).getExtensions(key);
         if (defaultFilters != null && defaultFilters.size() > 0) {
             filters.addAll(defaultFilters);
@@ -216,9 +217,28 @@ public class ProtocolFilterDecorator implements Protocol {
         // add filters via "filter" config
         String filterStr = url.getParameter(URLParamType.filter.getName());
         if (StringUtils.isNotBlank(filterStr)) {
-            String[] filterNames = MotanConstants.COMMA_SPLIT_PATTERN.split(filterStr);
+            HashSet<String> removedFilters = new HashSet<>();
+            Set<String> filterNames = StringTools.splitSet(filterStr, MotanConstants.COMMA_SEPARATOR);
             for (String fn : filterNames) {
-                addIfAbsent(filters, fn);
+                if (fn.startsWith(DISABLE_FILTER_PREFIX)) { // disable filter
+                    if (fn.length() > DISABLE_FILTER_PREFIX.length()) {
+                        removedFilters.add(fn.substring(DISABLE_FILTER_PREFIX.length()).trim());
+                    }
+                } else {
+                    Filter extFilter = ExtensionLoader.getExtensionLoader(Filter.class).getExtension(fn, false);
+                    if (extFilter == null) {
+                        LoggerUtil.warn("filter extension not found. filer name: " + fn);
+                        continue;
+                    }
+                    filters.add(extFilter);
+                }
+            }
+
+            // remove disabled filters
+            if (!removedFilters.isEmpty()) {
+                for (String removedName : removedFilters) {
+                    filters.removeIf((filter) -> removedName.equals(filter.getClass().getAnnotation(SpiMeta.class).name()));
+                }
             }
         }
 
@@ -226,29 +246,5 @@ public class ProtocolFilterDecorator implements Protocol {
         Collections.sort(filters, new ActivationComparator<>());
         Collections.reverse(filters);
         return filters;
-    }
-
-    private void addIfAbsent(List<Filter> filters, String extensionName) {
-        if (StringUtils.isBlank(extensionName)) {
-            return;
-        }
-
-        Filter extFilter = ExtensionLoader.getExtensionLoader(Filter.class).getExtension(extensionName, false);
-        if (extFilter == null) {
-            LoggerUtil.warn("filter extension not found. filer name: " + extensionName);
-            return;
-        }
-
-        boolean exists = false;
-        for (Filter f : filters) {
-            if (f.getClass() == extFilter.getClass()) {
-                exists = true;
-                break;
-            }
-        }
-        if (!exists) {
-            filters.add(extFilter);
-        }
-
     }
 }