+++ motan-core/src/main/java/com/weibo/api/motan/util/StringTools.java
@@ -19,15 +19,13 @@ package com.weibo.api.motan.util;
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 import java.net.URLEncoder;
-import java.util.Map;
-import java.util.TreeMap;
+import java.util.*;
 
 import org.apache.commons.lang3.StringUtils;
 
 import com.weibo.api.motan.common.MotanConstants;
 
 /**
- * 
  * String utils
  *
  * @author fishermen
@@ -72,7 +70,7 @@ public class StringTools {
     public static String toQueryString(Map<String, String> ps) {
         StringBuilder buf = new StringBuilder();
         if (ps != null && ps.size() > 0) {
-            for (Map.Entry<String, String> entry : new TreeMap<String, String>(ps).entrySet()) {
+            for (Map.Entry<String, String> entry : new TreeMap<>(ps).entrySet()) {
                 String key = entry.getKey();
                 String value = entry.getValue();
                 if (key != null && key.length() > 0 && value != null && value.length() > 0) {
@@ -86,5 +84,19 @@ public class StringTools {
             }
         }
         return buf.toString();
+    }
+
+    // 切分string，去重去空
+    public static Set<String> splitSet(String str, String regex) {
+        Set<String> result = new HashSet<>();
+        if (str != null) {
+            String[] strings = str.split(regex);
+            for (String s : strings) {
+                if (StringUtils.isNotBlank(s)) {
+                    result.add(s.trim());
+                }
+            }
+        }
+        return result;
     }
 }